move membership administration to own controller
