+++ src/system/GroupsModule/Resources/views/MembershipAdministration/userlist.html.twig
@@ -4,7 +4,7 @@
         <td headers="hUserName">{{ user.uname }}</td>
         <td headers="hEmail">{{ user.email }}</td>
         <td headers="hActions" class="actions">
-            <a class="fa fa-fw fa-user-plus tooltips" href="{{ path('zikulagroupsmodule_admin_adduser', {uid:user.uid, gid:gid, csrfToken:csrfToken}) }}" title="{{ __('Add user to this group') }}"></a>
+            <a class="fa fa-fw fa-user-plus tooltips" href="{{ path('zikulagroupsmodule_membershipadministration_add', {uid:user.uid, gid:gid, csrfToken:csrfToken}) }}" title="{{ __('Add user to this group') }}"></a>
         </td>
     </tr>
 {% else %}
