+++ src/system/ZAuthModule/Resources/views/UserAdministration/list.html.twig
@@ -1,4 +1,4 @@
-{{ pageAddAsset('javascript', zasset('@ZikulaZAuthModule:js/Zikula.ZAuth.Admin.View.js')) }}
+{{ pageAddAsset('javascript', zasset('@ZikulaUsersModule:js/Zikula.Users.Admin.View.js')) }}
 {{ adminHeader() }}
 <h3>
     <span class="fa fa-list"></span>
@@ -8,7 +8,7 @@
 
 <div id="user-search-container" class="form hide">
     <div class="input-group">
-        <input placeholder="{{ __('Search') }}" id="user-search" class="form-control" size="25" maxlength="25" type="text" value="" />
+        <input placeholder="{{ __('Search') }}" id="user-search" class="form-control" size="25" maxlength="25" type="text" data-route="zikulazauthmodule_useradministration_getusersbyfragmentastable" value="" />
         <span class="input-group-addon"><i class="fa fa-times pointer" id="user-search-discard"></i></span>
     </div>
 </div>
